Move bookmarks files to follow consistent approach
